cleanup syntax and comments
enforcing 80 chars per line is essentially impossible in Swift with the
crazy long Apple framework style function names, but in general this
should be a bit cleaner now.

OCD, sure, but I find it to be helpful when learning a new language to
go out of my way more than usual to make it look clean. Forces me to
experiment with different syntaxes to see what works and whatnot.