first stab at parsing git info
just rely on parsing output from git binary to try to do this in a zero
dependency way.

with this change we are starting to get closer to being a full (but
minimal) lolcommits replacement, which wasn’t my original intention but
hey if the shoe fits…

this could probably use some more cleanup but I’m on a plane and it’s
going to land soon so better go ahead and commit in case of horrible
fiery death and disaster.