Configuration management system
Switch to a configuration management system where all config values can
be set by either git config flags or environment variables.  This
enables the desired hierarchy of preferences.

Currently the class inheritence here results in some messy boilerplate,
so need to look into how to make that more elegant in the future.

As part of this change, --debug needs to be treated as a global flag so
it can affect everything even outside of the Configuration system.