convert from using path to NSURL in most places
This seems to be “in transition” in Swift/Cocoa, in a very annoying
way.  working with NSURL still requires a lot of the path methods from
NSString, and since they are no longer bridged to String there is a lot
of taking Strings (that were bridged from NSString to begin with) and
bridging them right back to NSString to get access to those methods.

The cleanest thing to do would be to just bridge all the path methods
back into String, but since I suspect they will eventually go away so
start migrating to the extra verbiage of dealing with NSURLs, which
makes me sad because it’s moving further out of the Swift STDLIB and
back into framework classes.  Sigh.

This also ends up with having to use NSURL.path! a lot, which I
*really* don’t like… but I see no sane alternative yet.